refactor: move api to api module, add .env variables
