+++ .env.example
@@ -1,3 +1,4 @@
 REDIS_ADDR=127.0.0.1:6379
 REDIS_PASSWORD=
 KAFKA=localhost:9092
+SOLIDITY_FULL_NODE_URL=
